refactor: detalhes de refatoração
