+++ pages/_app.public.js
@@ -1,6 +1,7 @@
 // _app.js
 import "/styles/globals.css";
 
+
 import { MqttProvider } from "./mqtt/MqttContext";
 import { mqttConfig } from "infra/mqttConfig";
 import { ThemeProvider } from "./components/ThemeToggle";
@@ -8,6 +9,7 @@ import { ThemeProvider } from "./components/ThemeToggle";
 export default function App({ Component, pageProps }) {
 
   return (
+
     <MqttProvider clientId={mqttConfig.clientId}>
       <ThemeProvider attribute="class"
         defaultTheme="system"
@@ -17,5 +19,7 @@ export default function App({ Component, pageProps }) {
       </ThemeProvider>
 
     </MqttProvider>
+
+
   );
 }
